minor fixes

+++ Application/Supplements/Commands/DeleteSupplement/DeleteSupplementCommand.cs
@@ -1,8 +1,10 @@
 ﻿using Application.Common;
+using Application.Common.Exceptions;
 using Application.Common.Interfaces;
 using Application.Common.Models;
 using Domain.Entities;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Caching.Memory;
 using System;
 using System.Threading;
@@ -43,6 +45,11 @@ namespace Application.Supplements.Commands.DeleteSupplement
 
                 if (supplement is null)
                     return Unit.Value;
+
+                var inUse = await db.Points.AnyAsync(p => p.SupplementId == request.SupplementId, cancellationToken: cancellationToken);
+
+                if (inUse)
+                    throw new BadRequestException("Засіб використовується.");
 
                 db.Supplements.Remove(supplement);
                 await db.SaveChangesAsync(cancellationToken);
